Removes one or more series from an rwl-style data frame

Removes one or more tree-ring series (columns) from a data frame of class "rwl". Optionally adds the rownames (assumed to be years) as a column.

Usage

trs_remove(x, series, rownames_to_years = FALSE, trim = FALSE)

Arguments

x

A data.frame of class "rwl", where each column is a tree-ring width series and rownames represent years.

series

A character vector specifying the names of the series (columns) to remove.

rownames_to_years

Logical. If TRUE, the rownames of x (years) are added as a column named year.

trim

Logical. If TRUE, leading and trailing all-NA rows are removed using trs_trim().

Value

A data.frame with the specified series removed. If rownames_to_years = TRUE, the result will have an additional year column.

Examples

rwl <- trs_pseudo_rwl(n_series = 3, series_length = 60, end_date = 2000)
trs_remove(rwl, series = "trs_1")
#>      trs_2 trs_3
#> 1941    93    95
#> 1942   102    95
#> 1943   113   101
#> 1944    97    89
#> 1945   109    99
#> 1946   113   103
#> 1947    98    82
#> 1948   105    90
#> 1949   106    98
#> 1950   106    99
#> 1951   113    99
#> 1952   112   113
#> 1953   121   119
#> 1954   106   121
#> 1955   109   121
#> 1956    96   110
#> 1957    84   110
#> 1958    73    97
#> 1959    71    88
#> 1960    79    98
#> 1961    89   103
#> 1962    91   104
#> 1963   101   111
#> 1964   111   113
#> 1965    96    93
#> 1966   114   104
#> 1967   111   102
#> 1968    93    90
#> 1969    91    94
#> 1970    86    93
#> 1971    86    89
#> 1972    89    90
#> 1973    82    87
#> 1974    89    93
#> 1975    86    88
#> 1976    82    87
#> 1977   102   101
#> 1978   103   107
#> 1979   108   111
#> 1980    90    99
#> 1981   112   112
#> 1982   110   113
#> 1983   106    97
#> 1984   116   115
#> 1985   112    98
#> 1986   112    99
#> 1987   103    92
#> 1988    98    81
#> 1989    92    72
#> 1990    80    68
#> 1991    86    81
#> 1992   106    97
#> 1993   105    99
#> 1994   106    98
#> 1995   105   108
#> 1996   114   112
#> 1997   109   101
#> 1998   119    99
#> 1999   125   110
#> 2000   120   111
trs_remove(rwl, series = c("trs_1", "trs_3"), rownames_to_years = TRUE)
#>      trs_2 year
#> 1941    93 1941
#> 1942   102 1942
#> 1943   113 1943
#> 1944    97 1944
#> 1945   109 1945
#> 1946   113 1946
#> 1947    98 1947
#> 1948   105 1948
#> 1949   106 1949
#> 1950   106 1950
#> 1951   113 1951
#> 1952   112 1952
#> 1953   121 1953
#> 1954   106 1954
#> 1955   109 1955
#> 1956    96 1956
#> 1957    84 1957
#> 1958    73 1958
#> 1959    71 1959
#> 1960    79 1960
#> 1961    89 1961
#> 1962    91 1962
#> 1963   101 1963
#> 1964   111 1964
#> 1965    96 1965
#> 1966   114 1966
#> 1967   111 1967
#> 1968    93 1968
#> 1969    91 1969
#> 1970    86 1970
#> 1971    86 1971
#> 1972    89 1972
#> 1973    82 1973
#> 1974    89 1974
#> 1975    86 1975
#> 1976    82 1976
#> 1977   102 1977
#> 1978   103 1978
#> 1979   108 1979
#> 1980    90 1980
#> 1981   112 1981
#> 1982   110 1982
#> 1983   106 1983
#> 1984   116 1984
#> 1985   112 1985
#> 1986   112 1986
#> 1987   103 1987
#> 1988    98 1988
#> 1989    92 1989
#> 1990    80 1990
#> 1991    86 1991
#> 1992   106 1992
#> 1993   105 1993
#> 1994   106 1994
#> 1995   105 1995
#> 1996   114 1996
#> 1997   109 1997
#> 1998   119 1998
#> 1999   125 1999
#> 2000   120 2000