Calculate correlation values between tree-ring series

This function calculates common correlation values between dated tree-ring series (x) and a set of reference chronologies (y). When no master chronologies are provided, each series in x is compared to all other series in x.

Only values are reported for pairs of series with a common overlap >= min_overlap.

The correlation values computed include:

glk: 'Gleichläufigkeit' or 'percentage of parallel variation' (Buras & Wilmking 2015; Eckstein & Bauch 1969; Huber 1942; Visser 2020)

glk_p: significance level associated with the glk-value (Jansma 1995)

r_pearson: the Pearson's correlation coefficient

t_St: Student's t-value

t_BP: t-values according to the Baillie & Pilcher (1973) algorithm

t_Ho: t-values according to the Hollstein (1980) algorithm

Usage

cor_table(
  x,
  y = NULL,
  min_overlap = 50,
  remove_duplicates = TRUE,
  output = "table",
  sort_by = "t_Ho"
)

Arguments

x

A data.frame of class ´rwl' with tree-ring data. Each column represents a measurement series, and row names correspond to (calendar) years.

y

A data.frame of class 'rwl' with tree-ring data. Each column represents a measurement series or chronology, and row names correspond to (calendar) years. If NULL (default), x is compared to itself (y = x).

min_overlap

A numeric value specifying the minimum overlap required between series for correlation calculation.

remove_duplicates

A logical value. If TRUE, identical pairs of series and references are removed from the output.

output

The desired output format, either "matrix" or "table" (default).

sort_by

The correlation value by which the output is sorted for each series in x. One of "r_pearson", "t_St", "glk", "glk_p", "t_BP", "t_Ho". Default to "t_Ho"

Value

The function returns a list of correlation matrices if output is set to "matrix." If output is set to "table," it returns a data.frame

reporting all correlation values.

Details

This function computes various correlation values between tree-ring series in a data.frame x and a set of reference chronologies in a data.frame y. If y is not provided, it compares each series in x to all other series in `x.

References

Baillie, M.G.L., Pilcher, J.R. (1973) A simple crossdating program for tree-ring research. Tree-Ring Bulletin 33, 7–14. http://hdl.handle.net/10150/260029

Buras, A. and Wilmking, M. (2015) Correcting the calculation of Gleichläufigkeit, Dendrochronologia 34, 29-30. https://doi.org/10.1016/j.dendro.2015.03.003

Eckstein, D. and Bauch, J. (1969) Beitrag zur Rationalisierung eines dendrochronologischen Verfahrens und zur Analyse seiner Aussagesicherheit. Forstwissenschaftliches Centralblatt, 88(1), 230-250.

Huber, B. (1943) Über die Sicherheit jahrringchronologischer Datierung. Holz als Roh- und Werkstoff 6, 263-268. https://doi.org/10.1007/BF02603303

Hollstein E. (1980) Mitteleuropäische Eichenchronologie. Trierer dendrochronologische Forschungen zur Archäologie und Kunstgeschichte, Trierer Grabungen und Forschungen 11, Mainz am Rhein.

Jansma, E. (1995) RemembeRINGs; The development and application of local and regional tree-ring chronologies of oak for the purposes of archaeological and historical research in the Netherlands, Nederlandse Archeologische Rapporten 19, Rijksdienst voor het Oudheidkundig Bodemonderzoek, Amersfoort. https://dspace.library.uu.nl/handle/1874/45149

Schweingruber, F. H. (1988) Tree rings: basics and applications of dendrochronology, Kluwer Academic Publishers, Dordrecht, Netherlands, 276 p.

Visser, R.M. (2020) On the similarity of tree-ring patterns: Assessing the influence of semi-synchronous growth changes on the Gleichläufigkeit for big tree-ring data sets, Archaeometry 63, 204-215. https://doi.org/10.1111/arcm.12600

Examples

# example code

Doel1 <- system.file("extdata", "DOEL1.fh", package = "fellingdater")
Doel1_trs <- read_fh(Doel1, verbose = FALSE)
# crossdating ring-width series from Doel 1 against each other:

cor_results <- cor_table(Doel1_trs, output = "table", min_overlap = 80,
sort_by = "t_Ho", remove_duplicates = TRUE)
head(cor_results,20)
#>     series length first last reference ref_first ref_last overlap  glk
#> 29 GD3-1BB     89  1222 1310    GQ1mBB      1150     1314      89 52.8
#> 73 GD3-1BB     89  1222 1310    S13mSB      1164     1322      89 48.9
#> 91 GD3-1BB     89  1222 1310     S6-SB      1221     1324      89 54.5
#> 58 GD3-1BB     89  1222 1310    K1_095      1207     1320      89 51.7
#> 31 GD3-1BB     89  1222 1310    GR1mBB      1220     1310      89 47.2
#> 87 GD3-1BB     89  1222 1310    S38-BB      1193     1306      85 54.2
#> 75 GG1-1BB     83  1240 1322    S13mSB      1164     1322      83 59.8
#> 69 GG1-1BB     83  1240 1322   S13A-BB      1232     1324      83 55.5
#> 59 GG1-1BB     83  1240 1322    K1_095      1207     1320      81 50.6
#> 98 GG1-1BB     83  1240 1322     S6-SB      1221     1324      83 47.0
#> 33  GQ1mBB    165  1150 1314    GR1mBB      1220     1310      91 76.7
#> 76  GQ1mBB    165  1150 1314    S13mSB      1164     1322     151 59.0
#> 60  GQ1mBB    165  1150 1314    K1_095      1207     1320     108 56.1
#> 47  GQ1mBB    165  1150 1314    K1_091      1158     1292     135 60.1
#> 62  GQ1mBB    165  1150 1314   S13A-BB      1232     1324      83 54.9
#> 89  GQ1mBB    165  1150 1314    S38-BB      1193     1306     114 49.6
#> 95  GQ1mBB    165  1150 1314     S6-SB      1221     1324      94 54.3
#> 77  GR1mBB     91  1220 1310    S13mSB      1164     1322      91 65.0
#> 51  GR1mBB     91  1220 1310    K1_095      1207     1320      91 60.0
#> 92  GR1mBB     91  1220 1310     S6-SB      1221     1324      90 55.6
#>           glk_p    r_pearson        t_St  t_BP  t_Ho
#> 29 5.919436e-01 -0.249779208 -2.40605037  2.11  2.74
#> 73 1.169772e+00  0.201284974  1.91669069  1.48  1.97
#> 91 3.910942e-01  0.035883266  0.33491251  2.16  1.92
#> 58 7.477459e-01  0.258187935  2.49273359  0.78  1.71
#> 31 1.408056e+00 -0.370271866 -3.71792327  0.78  1.18
#> 87 4.423117e-01  0.426532081  4.29630804  0.20  0.00
#> 75 7.546208e-02 -0.005084336 -0.04575962  2.28  2.93
#> 69 3.173465e-01  0.217882020  2.00920920  1.06  2.06
#> 59 9.104270e-01  0.597953189  6.63071551  0.09  1.07
#> 98 1.421457e+00 -0.137136419 -1.24599977  0.27  0.78
#> 33 3.624784e-07  0.727924641 10.01554116 12.23 10.41
#> 76 2.697516e-02  0.572471464  8.52261829  5.84  5.40
#> 60 2.067270e-01 -0.282768827 -3.03515368  3.52  3.02
#> 47 1.922538e-02  0.151940473  1.77284639  2.28  2.67
#> 62 3.740983e-01 -0.041724848 -0.37585094  2.19  2.48
#> 89 1.075278e+00 -0.178020527 -1.91457410  2.74  2.34
#> 95 4.042757e-01  0.078494804  0.75522595  1.09  1.16
#> 77 4.212192e-03  0.182334033  1.74946269  4.02  4.24
#> 51 5.640693e-02 -0.248806924 -2.42344980  3.38  3.50
#> 92 2.864524e-01  0.286284295  2.80290094  0.80  0.91