Report felling dates of individual tree-ring series

This function reports the felling date estimate of individual tree-ring series, based on the presence/absence of sapwood and/or waney edge. There are three possible modes of reporting:

a terminus post quem or earliest possible felling date: when only heartwood rings have been observed and measured

a felling date range or interval: when sapwood rings have been recorded, but no bark or waney edge is present.

an exact felling date: when bark or waney edge is present on the measured sample.

Reports the lower and upper boundaries of a felling date range for individual tree-ring series.

Usage

fd_report(
  x,
  series = "series",
  last = "last",
  n_sapwood = "n_sapwood",
  waneyedge = "waneyedge",
  sw_data = "Hollstein_1980",
  cred_mass = 0.954,
  densfun = "lognormal"
)

Arguments

x

Name of a data.frame with at least four columms, providing information on

the id's of the tree-ring series

the number of sapwood rings observed

the presence of waney edge

the date assigned to the last measured ring.

A column describing the sapwood data set to be used for modelling and the computation of the hdi can be provided as well.

series

Name of the column in x where id's of the tree-ring series are listed as character values.

last

Name of the column in x which lists the calendar year assigned to the last measured ring (should be a numeric vector).

n_sapwood

Name of the column in x where the number of observed sapwood rings are listed (should be numeric vector).

waneyedge

Name of the column in x indicating the presence (TRUE)/absence (FALSE) of waney edge (should be a logical vector).

sw_data

There are two options:

A character string providing the name of the sapwood data set to use for modelling. It should be one of the data sets listed in sw_data_overview(), or the name of a data.frame with sapwood data in columns n_sapwood and count, or

the name of the column in xthat lists for each series one of the sapwood data sets given by sw_data_overview().

cred_mass

A scalar [0, 1] specifying the mass within the credible interval (default = .954).

densfun

Name of the density function fitted to the sapwood data set. Should be one of:

lognormal (the default value),

normal,

weibull,

gammma.

Value

A data.frame reporting the estimate of the felling date for each tree-ring series?

See also

Examples

tmp <- data.frame(id = c("aaa", "bbb", "ccc"),
                  swr = c(10, 11, 12),
                  waneyedge = c(FALSE, FALSE,TRUE),
                  end = c(10, 0, -10))
fd_report(tmp,
          series = "id",
          n_sapwood = "swr",
          last = "end",
          sw_data = "Wazny_1990")
#>   series last n_sapwood waneyedge lower upper      felling_date sapwood_model
#> 1    aaa   10        10     FALSE    10    26 between 10 and 26    Wazny_1990
#> 2    bbb    0        11     FALSE     0    15  between 0 and 15    Wazny_1990
#> 3    ccc  -10        12      TRUE    NA   -10            in -10    Wazny_1990

# Example with different sw_model for individual series

sw_models_for_indiv_series <- c("Sohar_2012_ELL_c",
                                "Wazny_1990",
                                "Hollstein_1980",
                                "vanDaalen_Norway",
                                "vanDaalen_Norway")

trs_example2_edit <- cbind(trs_example2, "sw_models" = sw_models_for_indiv_series)

fd_report(trs_example2_edit,
          sw_data = "sw_models"
          )
#>   series last n_sapwood waneyedge lower upper          felling_date
#> 1 trs_06 1000         5     FALSE  1002  1013 between 1002 and 1013
#> 2 trs_07 1005        10     FALSE  1005  1021 between 1005 and 1021
#> 3 trs_08 1008        NA     FALSE  1014    NA            after 1014
#> 4 trs_09 1000         1     FALSE  1009  1031 between 1009 and 1031
#> 5 trs_10 1010        15      TRUE    NA  1010               in 1010
#>      sapwood_model
#> 1 Sohar_2012_ELL_c
#> 2       Wazny_1990
#> 3   Hollstein_1980
#> 4 vanDaalen_Norway
#> 5 vanDaalen_Norway