Compute a running mean on a time series

This vectorized function computes a running mean/moving average over a continuous (time) series with a specified window width.

Usage

movAv(x, w = 11, align = "center", edges = "fill")

Arguments

x

A numeric vector of length(x), containing the time series data.

w

The width of the moving average window. When wis even, one more value from the future is included.

align

The alignment of the window relative to the current data point. It can be one of:

align = "center": The average is assigned to the center of the window (default).

align = "left": The average includes the current value and the next (w-1) values.

align = "right": The average includes the current value and the previous (w-1) values.

edges

Defines how values are computed at the start and end of the series. Options are:

edges = "fill": The average is calculated with a decreasing number of values near the edges of the vector (default).

edges = "nofill": The result includes NA values when the window does not cover w values.

Value

A numeric vector of the same length of x with the computed running mean values.

Examples

num_vec <- rnorm(100)
movAv(num_vec, w = 20)
#>                [,1]
#>   [1,] -0.539839357
#>   [2,] -0.484237350
#>   [3,] -0.473831437
#>   [4,] -0.446533124
#>   [5,] -0.423480462
#>   [6,] -0.336993591
#>   [7,] -0.318429623
#>   [8,] -0.300175438
#>   [9,] -0.268448915
#>  [10,] -0.329552385
#>  [11,] -0.353959593
#>  [12,] -0.398457659
#>  [13,] -0.338341424
#>  [14,] -0.197214053
#>  [15,] -0.147313990
#>  [16,] -0.284864346
#>  [17,] -0.147554401
#>  [18,]  0.057775352
#>  [19,] -0.002662332
#>  [20,]  0.032366783
#>  [21,]  0.040108640
#>  [22,]  0.013993789
#>  [23,]  0.027185753
#>  [24,]  0.047067919
#>  [25,] -0.022278077
#>  [26,] -0.066346051
#>  [27,] -0.076882515
#>  [28,] -0.077754490
#>  [29,] -0.053147177
#>  [30,]  0.088285333
#>  [31,]  0.036068008
#>  [32,]  0.129018765
#>  [33,]  0.080367772
#>  [34,]  0.044225222
#>  [35,]  0.103849269
#>  [36,]  0.207131550
#>  [37,]  0.184704436
#>  [38,]  0.116798020
#>  [39,]  0.106764890
#>  [40,]  0.152972278
#>  [41,]  0.160678935
#>  [42,]  0.149953691
#>  [43,]  0.112539850
#>  [44,]  0.070349133
#>  [45,]  0.130255962
#>  [46,]  0.085572296
#>  [47,]  0.044998682
#>  [48,]  0.003135202
#>  [49,] -0.006890511
#>  [50,] -0.122905614
#>  [51,] -0.066912366
#>  [52,] -0.180136098
#>  [53,] -0.090600409
#>  [54,] -0.138758143
#>  [55,] -0.215315844
#>  [56,] -0.290683579
#>  [57,] -0.431710374
#>  [58,] -0.539822939
#>  [59,] -0.409031618
#>  [60,] -0.450331434
#>  [61,] -0.460983988
#>  [62,] -0.442898218
#>  [63,] -0.402427012
#>  [64,] -0.365662144
#>  [65,] -0.444412722
#>  [66,] -0.433036970
#>  [67,] -0.422432192
#>  [68,] -0.381374424
#>  [69,] -0.428352720
#>  [70,] -0.412238785
#>  [71,] -0.431181536
#>  [72,] -0.411402258
#>  [73,] -0.514648110
#>  [74,] -0.482782631
#>  [75,] -0.555698845
#>  [76,] -0.398443040
#>  [77,] -0.246931114
#>  [78,] -0.147112595
#>  [79,] -0.268911019
#>  [80,] -0.191038104
#>  [81,] -0.160903431
#>  [82,] -0.094882153
#>  [83,] -0.112462698
#>  [84,] -0.091079136
#>  [85,]  0.032339438
#>  [86,]  0.003738585
#>  [87,]  0.027708262
#>  [88,]  0.068980303
#>  [89,]  0.046276319
#>  [90,]  0.073515208
#>  [91,]  0.045744473
#>  [92,]  0.095699346
#>  [93,]  0.171039002
#>  [94,]  0.203827338
#>  [95,]  0.307537129
#>  [96,]  0.215679333
#>  [97,]  0.173878152
#>  [98,]  0.134107898
#>  [99,]  0.213762790
#> [100,]  0.131901837