Select series from an rwl-style data frame

Extracts one or more tree-ring series (columns) from a data frame of class "rwl". Optionally adds the rownames (assumed to be years) as a column.

Usage

trs_select(x, series, rownames_to_years = FALSE, trim = FALSE)

Arguments

x

A data.frame of class "rwl", where each column is a tree-ring width series and rownames represent years.

series

A character string or vector specifying the name(s) of the series (columns) to extract.

rownames_to_years

Logical. If TRUE, the rownames of x (years) are added as a column named years.

trim

Logical. If TRUE, leading and trailing all-NA rows are removed using trs_trim().

Value

A data.frame with the selected series. If rownames_to_years = TRUE, the result will have an additional year column. If trim = TRUE, leading and trailing all-NA rows are removed. If a single series is selected, returns a single-column data frame.

Examples

rwl <- trs_pseudo_rwl(n_series = 5, series_length = 60, end_date = 2000)

# Select single series
trs_select(rwl, series = "trs_1")
#>      trs_1
#> 1941   103
#> 1942   105
#> 1943   114
#> 1944   103
#> 1945   103
#> 1946   103
#> 1947   113
#> 1948   112
#> 1949   112
#> 1950    96
#> 1951   109
#> 1952   107
#> 1953   107
#> 1954   115
#> 1955   108
#> 1956   100
#> 1957   108
#> 1958   101
#> 1959   100
#> 1960    98
#> 1961    94
#> 1962   102
#> 1963   115
#> 1964   108
#> 1965   113
#> 1966   109
#> 1967   113
#> 1968   101
#> 1969   108
#> 1970   102
#> 1971   116
#> 1972    94
#> 1973    88
#> 1974    91
#> 1975    72
#> 1976    72
#> 1977    76
#> 1978    83
#> 1979    79
#> 1980    90
#> 1981    93
#> 1982   101
#> 1983   105
#> 1984   104
#> 1985   114
#> 1986   105
#> 1987    98
#> 1988   116
#> 1989   115
#> 1990   118
#> 1991   107
#> 1992   110
#> 1993    91
#> 1994    70
#> 1995    93
#> 1996    92
#> 1997    94
#> 1998    94
#> 1999    84
#> 2000    89
trs_select(rwl, series = "trs_2", rownames_to_years = TRUE)
#>      trs_2 year
#> 1941    98 1941
#> 1942   100 1942
#> 1943   104 1943
#> 1944    90 1944
#> 1945    93 1945
#> 1946    96 1946
#> 1947   104 1947
#> 1948   106 1948
#> 1949   102 1949
#> 1950    91 1950
#> 1951   100 1951
#> 1952   103 1952
#> 1953    98 1953
#> 1954   104 1954
#> 1955   109 1955
#> 1956   101 1956
#> 1957   103 1957
#> 1958   100 1958
#> 1959    92 1959
#> 1960    88 1960
#> 1961    84 1961
#> 1962    94 1962
#> 1963   105 1963
#> 1964   101 1964
#> 1965   102 1965
#> 1966   100 1966
#> 1967   107 1967
#> 1968    95 1968
#> 1969   103 1969
#> 1970   104 1970
#> 1971   115 1971
#> 1972    97 1972
#> 1973    94 1973
#> 1974    87 1974
#> 1975    70 1975
#> 1976    77 1976
#> 1977    78 1977
#> 1978    82 1978
#> 1979    81 1979
#> 1980    95 1980
#> 1981    92 1981
#> 1982   104 1982
#> 1983   114 1983
#> 1984   112 1984
#> 1985   124 1985
#> 1986   110 1986
#> 1987   102 1987
#> 1988   123 1988
#> 1989   125 1989
#> 1990   125 1990
#> 1991   120 1991
#> 1992   117 1992
#> 1993    93 1993
#> 1994    75 1994
#> 1995    89 1995
#> 1996    83 1996
#> 1997    77 1997
#> 1998    92 1998
#> 1999    93 1999
#> 2000    94 2000

# Select multiple series
trs_select(rwl, series = c("trs_1", "trs_2"))
#>      trs_1 trs_2
#> 1941   103    98
#> 1942   105   100
#> 1943   114   104
#> 1944   103    90
#> 1945   103    93
#> 1946   103    96
#> 1947   113   104
#> 1948   112   106
#> 1949   112   102
#> 1950    96    91
#> 1951   109   100
#> 1952   107   103
#> 1953   107    98
#> 1954   115   104
#> 1955   108   109
#> 1956   100   101
#> 1957   108   103
#> 1958   101   100
#> 1959   100    92
#> 1960    98    88
#> 1961    94    84
#> 1962   102    94
#> 1963   115   105
#> 1964   108   101
#> 1965   113   102
#> 1966   109   100
#> 1967   113   107
#> 1968   101    95
#> 1969   108   103
#> 1970   102   104
#> 1971   116   115
#> 1972    94    97
#> 1973    88    94
#> 1974    91    87
#> 1975    72    70
#> 1976    72    77
#> 1977    76    78
#> 1978    83    82
#> 1979    79    81
#> 1980    90    95
#> 1981    93    92
#> 1982   101   104
#> 1983   105   114
#> 1984   104   112
#> 1985   114   124
#> 1986   105   110
#> 1987    98   102
#> 1988   116   123
#> 1989   115   125
#> 1990   118   125
#> 1991   107   120
#> 1992   110   117
#> 1993    91    93
#> 1994    70    75
#> 1995    93    89
#> 1996    92    83
#> 1997    94    77
#> 1998    94    92
#> 1999    84    93
#> 2000    89    94
trs_select(rwl, series = c("trs_1", "trs_3", "trs_5"), rownames_to_years = TRUE)
#>      trs_1 trs_3 trs_5 year
#> 1941   103   104   105 1941
#> 1942   105   114   105 1942
#> 1943   114   124   107 1943
#> 1944   103   109    97 1944
#> 1945   103   104   100 1945
#> 1946   103   111   105 1946
#> 1947   113   117   108 1947
#> 1948   112   114   112 1948
#> 1949   112   121   114 1949
#> 1950    96   100   105 1950
#> 1951   109   105   103 1951
#> 1952   107   101    98 1952
#> 1953   107    99    89 1953
#> 1954   115   108   103 1954
#> 1955   108   111   101 1955
#> 1956   100   101    98 1956
#> 1957   108   105   108 1957
#> 1958   101    98   104 1958
#> 1959   100   103   106 1959
#> 1960    98   103   102 1960
#> 1961    94    90    94 1961
#> 1962   102   104   105 1962
#> 1963   115   112   113 1963
#> 1964   108   108   110 1964
#> 1965   113   109   112 1965
#> 1966   109   104   111 1966
#> 1967   113   104   114 1967
#> 1968   101    92   104 1968
#> 1969   108   103   103 1969
#> 1970   102    96   109 1970
#> 1971   116   109   119 1971
#> 1972    94    94    97 1972
#> 1973    88    83   103 1973
#> 1974    91    80   101 1974
#> 1975    72    69    80 1975
#> 1976    72    70    77 1976
#> 1977    76    74    80 1977
#> 1978    83    84    90 1978
#> 1979    79    75    83 1979
#> 1980    90    82    96 1980
#> 1981    93    90   100 1981
#> 1982   101    98   102 1982
#> 1983   105   104   112 1983
#> 1984   104   103   109 1984
#> 1985   114   119   125 1985
#> 1986   105   117   114 1986
#> 1987    98   113   102 1987
#> 1988   116   126   121 1988
#> 1989   115   128   120 1989
#> 1990   118   128   118 1990
#> 1991   107   117   113 1991
#> 1992   110   116   112 1992
#> 1993    91    90    95 1993
#> 1994    70    77    79 1994
#> 1995    93    93    90 1995
#> 1996    92    88    93 1996
#> 1997    94    88    85 1997
#> 1998    94    98    89 1998
#> 1999    84    93    84 1999
#> 2000    89    92    90 2000